+++ init.lua
@@ -36,12 +36,12 @@ setmetatable(ui.dialogs, {
   end
 })
 textadept.file_types = {extensions = lexer.detect_extensions, patterns = lexer.detect_patterns}
-events.connect(events.VIEW_NEW, function()
-  for _, k in ipairs{'colors', 'styles'} do rawset(lexer, k, view[k]) end
-end)
+-- LuaFormatter off
+events.connect(events.VIEW_NEW, function() for _, k in ipairs{'colors', 'styles'} do rawset(lexer, k, view[k]) end end)
 setmetatable(lexer, {__newindex = function(_, k, v) if k:find('^fold') then view[k] = v end end})
 textadept.editing.INDIC_BRACEMATCH = _SCINTILLA.next_indic_number()
-textadept.editing.brace_matches = {}
+textadept.editing.brace_matches, textadept.editing.api_files = {}, setmetatable({}, {__index = function(t, k) t[k] = {} return t[k] end })
+-- LuaFormatter on
 
 -- The remainder of this file defines default buffer and view properties and applies them to
 -- subsequent buffers and views. Normally, a setting like `buffer.use_tabs = false` only applies